Create water_sensor.py
This is the code to run in the venv that will read the output of the arduino that it is listening for, assuming it is set up with 2 flow rate sensors and 4 temp sensors. The SENSOR_PREFIX can be changed to reflect whichever environment it is monitoring, and it will create all new config files for the new sensors.